requirements.txt
jupyter
python-swiftclient
python-keystoneclient
keystoneauth1
traitlets
requests
nose
mock
